Unhide song history pages for everyone
No reason they shouldn't be able to see song history, and in fact it may be helpful to find newest songs or check if a song has been modified since they last viewed it. And it can give a sense of more aliveness and activity to the app.
They always had access if they manipulated url, but now we unhide the links.
Also moved print button to very left of all buttons on the songs page for better organization and fixed its spacing.