Reconstruct user from cookie instead of loading from database (#165)
When I implemented support for user roles, I was wary of client-side tampering. Therefore, in my implementation, GT always read the users table for the user's role instead of storing and retrieving the role from the DB. Because every single action on GT requires checking the user's role, this forced a `select * from users` query for every single request.

However, I learned that Rails cookies are secure and tamper proof so this is completely unnecessary. To save unnecessary trips to the DB, this PR saves all the user's important fields in the cookie and reads them out of the cookie to re-construct the user. Turns out this only saves 1-4ms per request but it is fun to implement and makes GT feel more professional 💪.