Refresh user's persisted name during sign in (#174)
Currently, when a user signs in for the first time, we save their name as Google during the OAuth flow into the database...and never ever never update it after that, even when they sign in again. In other words, user names are fixed.

This is a minor inconvenience for anyone whose name changes after their initial sign in, which is very common case for sisters who get married. Let's respect people's name changes by refreshing the user's name with the name provided during OAuth every sign in. 

I also greatly improved test coverage of the sessions controller and renamed away all vestigial instances of "Gracepoint".